requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pandas>=2.0.0
plotly>=5.15.0
pyyaml>=6.0
jsonrpclib-pelix>=0.4.3
openpyxl>=3.1.0
werkzeug>=2.3.0
urllib3>=2.0.0